refactor: update update handling and modify banner labels for clarity
