+++ forge.default.yaml
@@ -2,7 +2,7 @@ variables:
   mode: ACT
 
 updates:
-  check_frequency: "weekly"
+  frequency: "weekly"
   auto_update: false
 
 # Define model anchors with simpler, purpose-based names
